+++ MGSwipeTableCell/MGSwipeButton.m
@@ -161,18 +161,20 @@
     return [[NSBundle mainBundle].executablePath rangeOfString:@".appex/"].location != NSNotFound;
 }
 
-
 -(BOOL) isRTLLocale
 {
+#if __IPHONE_OS_VERSION_MAX_ALLOWED >= 110000
+    if (@available(iOS 9, *)) {
+#else
     if ([[UIView class] respondsToSelector:@selector(userInterfaceLayoutDirectionForSemanticContentAttribute:)]) {
+#endif
         return [UIView userInterfaceLayoutDirectionForSemanticContentAttribute:self.semanticContentAttribute] == UIUserInterfaceLayoutDirectionRightToLeft;
+    }
+    else if ([self isAppExtension]) {
+        return [NSLocale characterDirectionForLanguage:[[NSLocale currentLocale] objectForKey:NSLocaleLanguageCode]]==NSLocaleLanguageDirectionRightToLeft;
     } else {
-        if ([self isAppExtension]) {
-            return [NSLocale characterDirectionForLanguage:[[NSLocale currentLocale] objectForKey:NSLocaleLanguageCode]]==NSLocaleLanguageDirectionRightToLeft;
-        } else {
-            UIApplication *application = [UIApplication performSelector:@selector(sharedApplication)];
-            return application.userInterfaceLayoutDirection == UIUserInterfaceLayoutDirectionRightToLeft;
-        }
+        UIApplication *application = [UIApplication performSelector:@selector(sharedApplication)];
+        return application.userInterfaceLayoutDirection == UIUserInterfaceLayoutDirectionRightToLeft;
     }
 }