+++ MGSwipeTableCell/MGSwipeTableCell.h
@@ -36,7 +36,7 @@ typedef NS_ENUM(NSInteger, MGSwipeState) {
     MGSwipeStateExpandingRightToLeft,
 };
 
-/** Swipe state */
+/** Swipe Expansion Layout */
 typedef NS_ENUM(NSInteger, MGSwipeExpansionLayout) {
     MGSwipeExpansionLayoutBorder = 0,
     MGSwipeExpansionLayoutCenter,
@@ -88,6 +88,8 @@ typedef NS_ENUM(NSInteger, MGSwipeEasingFunction) {
 @property (nonatomic, assign) CGFloat bottomMargin;
 /** Distance between the buttons. Default value : 0 */
 @property (nonatomic, assign) CGFloat buttonsDistance;
+/** If true, expands the last button length by safeAreaInsets. Useful for devices with a notch (e.g. iPhone X) */
+@property (nonatomic, assign) BOOL expandLastButtonBySafeAreaInsets;
 
 /** Animation settings when the swipe buttons are shown */
 @property (nonatomic, strong, nonnull) MGSwipeAnimation * showAnimation;
@@ -113,6 +115,9 @@ typedef NS_ENUM(NSInteger, MGSwipeEasingFunction) {
 /** Coefficient applied to cell movement in bounce zone. Set to value between 0.0 and 1.0
     to make the cell 'resist' swiping after buttons are revealed. Default is 1.0 */
 @property (nonatomic, assign) CGFloat swipeBounceRate;
+
+// default is NO. Controls whether buttons with different width are allowed. Buttons are resized to have the same size by default.
+@property (nonatomic) BOOL allowsButtonsWithDifferentWidth;
 
 @end
 
@@ -246,7 +251,7 @@ typedef NS_ENUM(NSInteger, MGSwipeEasingFunction) {
 // default is NO. Controls whether multiple cells can be swiped simultaneously
 @property (nonatomic) BOOL allowsMultipleSwipe;
 // default is NO. Controls whether buttons with different width are allowed. Buttons are resized to have the same size by default.
-@property (nonatomic) BOOL allowsButtonsWithDifferentWidth;
+@property (nonatomic) BOOL allowsButtonsWithDifferentWidth DEPRECATED_MSG_ATTRIBUTE("Use MGSwipeSettings.allowsButtonsWithDifferentWidth instead");
 //default is YES. Controls whether swipe gesture is allowed when the touch starts into the swiped buttons
 @property (nonatomic) BOOL allowsSwipeWhenTappingButtons;
 //default is YES. Controls whether swipe gesture is allowed in opposite directions. NO value disables swiping in opposite direction once started in one direction